requirements.txt
# Core LLM Evaluation Framework Dependencies
# Version pinning for production stability

# HuggingFace Ecosystem
transformers>=4.40.0,<5.0.0
sentence-transformers>=2.3.0,<3.0.0
datasets>=2.16.0,<3.0.0
accelerate>=0.27.0,<1.0.0

# Vector Search
faiss-cpu>=1.7.4,<2.0.0
scikit-learn>=1.4.0,<2.0.0

# Vector Databases
qdrant-client>=1.7.0,<2.0.0
chromadb>=0.4.0,<1.0.0

# PostgreSQL + pgvector
psycopg2-binary>=2.9.0,<3.0.0

# RAG Evaluation Framework
ragas>=0.1.0,<1.0.0

# AWS Integration
boto3>=1.34.0,<2.0.0
botocore>=1.34.0,<2.0.0

# Google Drive API (for accessing Colab notebooks)
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0

# OpenAI API
openai>=1.12.0,<2.0.0

# Statistical Analysis
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scipy>=1.12.0,<2.0.0

# Code Quality Analysis
radon>=6.0.0,<7.0.0
mccabe>=0.7.0,<1.0.0

# Testing Framework
pytest>=8.0.0,<9.0.0
pytest-cov>=4.1.0,<5.0.0

# Logging and Configuration
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# Jupyter Notebook Support
jupyter>=1.0.0,<2.0.0
ipykernel>=6.29.0,<7.0.0
notebook>=7.0.0,<8.0.0

# Type Checking
mypy>=1.8.0,<2.0.0
types-pyyaml>=6.0.0,<7.0.0

# Utilities
tqdm>=4.66.0,<5.0.0
rich>=13.7.0,<14.0.0

# Visualization
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

